requirements.txt
google-cloud-storage
google-cloud-logging
google-cloud-firestore
requests
